RemoveAdminPatchをDisableDevice.csに移動
